requirements.txt
Django==3.2.23
djangorestframework==3.14.0
psycopg2==2.8.0
